requirements.txt
Django==2.1
django-appconf==1.0.3
django-ckeditor==5.7.1
django-debug-toolbar==2.0
django-extensions==2.1.9
-e git+https://github.com/rodrigoamaral/django-fullcalendar.git@d116fcfc2093e57035932a9ea7c111d449fd7e4d#egg=django_fullcalendar
django-imagekit==4.0.2
django-js-asset==1.2.2
django-location-field==2.1.0
django-tagging==0.4.6
pilkit==2.0
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
